requirements.txt
streamlit-option-menu
ultralytics
fuzzywuzzy
flake8
pycodestyle_magic
inference_sdk
aiohttp
aiosignal
altair
async-timeout
attrs
blinker
cachetools
certifi
charset-normalizer
click
dataclasses-json
decorator
docx
faiss-cpu
frozenlist
gitdb
GitPython
googlemaps
geopy
greenlet
idna
importlib-metadata
install
Jinja2
jsonschema
langchain
langchainplus-sdk
lxml
Markdown
markdown-it-py
MarkupSafe
marshmallow
marshmallow-enum
mdurl
multidict
mypy-extensions
numexpr
numpy
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging
pandas
Pillow
protobuf
pyarrow
pydantic
pydeck
Pygments
Pympler
pyrsistent
python-dateutil
python-docx
python-dotenv
pytz
pytz-deprecation-shim
PyYAML
requests
rich
six
smmap
SQLAlchemy
streamlit
tenacity
tiktoken
toml
toolz
tornado
tqdm
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
validators
yarl
zipp
opencv-python
python-Levenshtein
langchain-community